refactor, dev: расширена структура бизнес-логики
