+++ src/handlers/logic.py
@@ -15,13 +15,15 @@ from src.database import (
     update_service,
 )
 from src.exceptions import ServiceNameTooLongError
-from src.business_logic.get_times_possible_for_appointment import get_times_possible_for_appointment
-from src.business_logic.utils import (
+from src.business_logic.make_appointment.get_times_possible_for_appointment import get_times_possible_for_appointment
+from src.business_logic.make_appointment.utils import (
     get_days_keyboard_buttons,
+    get_years_with_months_days,
+)
+from src.business_logic.resolve_times_statuses.utils import (
     get_initial_times_statuses,
     get_set_working_hours_keyboard_buttons,
     get_times_statuses_view,
-    get_years_with_months_days,
 )
 from src.keyboards import (
     BACK,
